Module Bolt.Type_async_expr

val type_async_expr : Ast_types.class_defn list -> Ast_types.trait_defn list -> Typed_ast.expr -> (Type_env.type_env, Core.Error.t) Core.Result.t

The "Ok" result type - type_env is not used at all when the function returns, it is only used when the function makes recursive calls to itself