Module Bolt.Type_linear_ownership

type ownership =

| LinearOwned

| LinearFree

| NonLinear

val type_linear_ownership : Ast_types.class_defn list -> Ast_types.trait_defn list -> Typed_ast.expr -> (ownership, Core.Error.t) Core.Result.t

The "Ok" result type ("") is not used at all when the function returns, it is only used when the function makes recursive calls to itself