Module Bolt__Eval_step

We use small-step semantics when executing the interpreter. This module takes in a state (thread pool and heap) and the scheduled thread's ID and then executes a single step of execution (using top instruction in the code stack of that thread, and returns the updated state if successful

val eval_step : Bolt.Runtime_env.thread_pool -> Bolt.Runtime_env.heap -> scheduled_thread_id:Bolt.Runtime_env.threadID -> (Bolt.Runtime_env.thread_pool * Bolt.Runtime_env.heap) Core.Or_error.t