Module Bolt__Pprint_ast

This module defines a set of pretty printing functions for the parts of the AST that are common across the parsed and typed ASTs - to be used by pretty printing functions of the two ASTs.

They all take in a formatter which specifies output channel e.g. a string formatter or a stdout formatter. (Think of this as a generalisation of printf)

The second argument for all of these is an indent - this corresponds to nesting depth within the AST

val pprint_class_defn : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.class_defn -> unit

val pprint_trait_defn : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.trait_defn -> unit

val pprint_capability : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.capability -> unit

val pprint_cap_trait : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.cap_trait -> unit

val pprint_mode : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.mode -> unit

val pprint_type_field : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.type_field -> unit

val pprint_field_defn : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.field_defn -> unit

val pprint_require_field_defn : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.require_field_defn -> unit

val pprint_type_expr : Stdlib.Format.formatter -> indent:string -> Bolt.Ast_types.type_expr -> unit