Module Bolt__Type_async_expr

This module checks thread references are not accessed from another thread

val type_async_expr : Bolt.Ast_types.class_defn list -> Bolt.Ast_types.trait_defn list -> Bolt.Typed_ast.expr -> (Bolt.Type_env.type_env, Core.Error.t) Core.Result.t

The "Ok" result type - type_env is not used at all when the function returns, it is only used when the function makes recursive calls to itself