Module Bolt__Type_checker

The type checker takes a Parsed AST and type-checks the core language rules, returning a typed AST if type-checking succeeds. It also optionally performs a second stage of type-checking to check for data-race freedom.

val type_check_program : Bolt.Parsed_ast.program -> check_data_races:bool -> Bolt.Typed_ast.program Core.Or_error.t

val pprint_typed_ast : Stdlib.Format.formatter -> Bolt.Typed_ast.program -> unit

Given a formatter and typed AST, pretty-print the AST - useful for debugging