Module Bolt__Typed_ast

The typed AST consists of the original AST but augmented with type information.

type expr =

| Integer of Bolt.Ast_types.loc * int
no need for type_expr annotation as obviously TEInt

| Variable of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * Bolt.Ast_types.Var_name.t

| Lambda of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * Bolt.Ast_types.Var_name.t * Bolt.Ast_types.type_expr * expr

| App of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * expr * expr

| Seq of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * expr list

| Let of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * Bolt.Ast_types.Var_name.t * expr * expr

| ObjField of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * Bolt.Ast_types.Var_name.t * Bolt.Ast_types.type_expr * Bolt.Ast_types.Field_name.t
First type is of the expr, second is the type of the obj

| Assign of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * Bolt.Ast_types.Var_name.t * Bolt.Ast_types.type_expr * Bolt.Ast_types.Field_name.t * expr
First type is of the expr, second is the type of the obj

| Constructor of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * Bolt.Ast_types.Class_name.t * constructor_arg list

| Consume of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * expr

| FinishAsync of Bolt.Ast_types.loc * Bolt.Ast_types.type_expr * expr * expr * expr

and constructor_arg =

| ConstructorArg of Bolt.Ast_types.type_expr * Bolt.Ast_types.Field_name.t * expr

type program =

| Prog of Bolt.Ast_types.class_defn list * Bolt.Ast_types.trait_defn list * expr