Module Hello_world

Hello World

Hello world is a module I'm using to learn fun Ocaml build and testing and documentation tools.

I've made this change to test automatic post commit hook!

val id : int -> int

id a returns the result a

val zero : int -> int

zero a returns the result 0, regardless of the value of a

val succ : int -> int

succ a returns the successor of a, i.e. a+1